Colocate signatures with classes
