+++ packages/components/src/components/hds/interactive/index.ts
@@ -5,7 +5,24 @@
 
 import Component from '@glimmer/component';
 import { action } from '@ember/object';
-import type { HdsInteractiveSignature } from './types';
+
+export interface HdsInteractiveSignature {
+  Args: {
+    href?: string;
+    isHrefExternal?: boolean;
+    isRouteExternal?: boolean;
+    route?: string;
+    models?: Array<string | number>;
+    model?: string | number;
+    query?: Record<string, string>;
+    'current-when'?: string;
+    replace?: boolean;
+  };
+  Blocks: {
+    default: [];
+  };
+  Element: HTMLAnchorElement | HTMLButtonElement;
+}
 
 export default class HdsInteractiveIndexComponent extends Component<HdsInteractiveSignature> {
   /**
