--- packages/components/src/components/hds/interactive/types.ts
@@ -1,22 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: MPL-2.0
- */
-
-export interface HdsInteractiveSignature {
-  Args: {
-    href?: string;
-    isHrefExternal?: boolean;
-    isRouteExternal?: boolean;
-    route?: string;
-    models?: Array<string | number>;
-    model?: string | number;
-    query?: Record<string, string>;
-    'current-when'?: string;
-    replace?: boolean;
-  };
-  Blocks: {
-    default: [];
-  };
-  Element: HTMLAnchorElement | HTMLButtonElement;
-}
