+++ packages/components/src/components/hds/link/types.ts
@@ -3,8 +3,6 @@
  * SPDX-License-Identifier: MPL-2.0
  */
 
-import type { HdsInteractiveSignature } from '../interactive/types.ts';
-
 export enum HdsLinkIconPositionValues {
   Leading = 'leading',
   Trailing = 'trailing',
@@ -26,32 +24,3 @@ export enum HdsLinkStandaloneSizeValues {
 }
 
 export type HdsLinkStandaloneSizes = `${HdsLinkStandaloneSizeValues}`;
-
-export interface HdsLinkStandaloneSignature {
-  Args: HdsInteractiveSignature['Args'] & {
-    icon: string;
-    text: string;
-    color?: HdsLinkColors;
-    href?: string;
-    iconPosition?: HdsLinkIconPositions;
-    isHrefExternal?: boolean;
-    isRouteExternal?: boolean;
-    size?: HdsLinkStandaloneSizes;
-  };
-  Element: HdsInteractiveSignature['Element'];
-}
-
-export interface HdsLinkInlineSignature {
-  Args: HdsInteractiveSignature['Args'] & {
-    color?: HdsLinkColors;
-    href?: string;
-    icon?: string;
-    iconPosition?: HdsLinkIconPositions;
-    isHrefExternal?: boolean;
-    isRouteExternal?: boolean;
-  };
-  Blocks: {
-    default: [];
-  };
-  Element: HdsInteractiveSignature['Element'];
-}
