--- packages/components/src/components/hds/yield/types.ts
@@ -1,10 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: MPL-2.0
- */
-
-export interface HdsYieldSignature {
-  Blocks: {
-    default: [];
-  };
-}
